backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
numpy==1.21.1
werkzeug==2.0.1
python-dotenv==0.19.0
python-clamd==1.0.2
albumentations==1.3.0 